Restructured code base clustering functionality in separate files
