+++ README.md
@@ -1,6 +1,6 @@
 # gosymbol
 
-<img align="right" width="159px" src="./gopher-doing-maths.jpg">
+<img align="right" width="159px" src="./assets/images/gopher-doing-maths.jpg">
 
 `gosymbol` is a Go package for performing symbolic computations. Its main purpose is to be educational and to give me a chance to dive into the inner workings of symbolic computations and computer algebra systems. The development journey will be published as a series of [Medium articles](https://medium.com/@victor.w.brun).
