+++ eval.go
@@ -1,0 +1,104 @@
+package gosymbol
+
+import (
+	"fmt"
+	"math"
+)
+
+/* Evaluation */
+
+func (e undefined) Eval() Func {
+	return func(args Arguments) float64 { return math.NaN() }
+}
+
+func (e constant) Eval() Func {
+	return func(args Arguments) float64 { return e.Value }
+}
+
+func (e variable) Eval() Func {
+	return func(args Arguments) float64 { return args[e.Name] }
+}
+
+func (e add) Eval() Func {
+	return func(args Arguments) float64 {
+		sum := e.Operands[0].Eval()(args) // Initiate with first operand since 0 may not always be identity
+		for ix := 1; ix < len(e.Operands); ix++ {
+			sum += e.Operands[ix].Eval()(args)
+		}
+		return sum
+	}
+}
+
+func (e mul) Eval() Func {
+	return func(args Arguments) float64 {
+		prod := e.Operands[0].Eval()(args) // Initiate with first operand since 1 may not always be identity
+		for ix := 1; ix < len(e.Operands); ix++ {
+			prod *= e.Operands[ix].Eval()(args)
+		}
+		return prod
+	}
+}
+
+func (e exp) Eval() Func {
+	return func(args Arguments) float64 { return math.Exp(e.Arg.Eval()(args)) }
+}
+
+func (e log) Eval() Func {
+	return func(args Arguments) float64 { return math.Log(e.Arg.Eval()(args)) }
+}
+
+func (e pow) Eval() Func {
+	return func(args Arguments) float64 { return math.Pow(e.Base.Eval()(args), e.Exponent.Eval()(args)) }
+}
+
+/* Evaluation to string */
+
+func (e undefined) String() string {
+	return "Undefined"
+}
+
+func (e constant) String() string {
+	if e.Value < 0 {
+		return fmt.Sprintf("( %v )", e.Value)
+	} else {
+		return fmt.Sprint(e.Value)
+	}
+}
+
+func (e variable) String() string {
+	return string(e.Name)
+}
+
+func (e constrainedVariable) String() string {
+	return fmt.Sprintf("%v_CONSTRAINED", e.Name)
+}
+
+func (e add) String() string {
+	str := fmt.Sprintf("( %v", e.Operands[0])
+	for ix := 1; ix < len(e.Operands); ix++ {
+		str += fmt.Sprintf(" + %v", e.Operands[ix])
+	}
+	str += " )"
+	return str
+}
+
+func (e mul) String() string {
+	str := fmt.Sprintf("( %v", e.Operands[0])
+	for ix := 1; ix < len(e.Operands); ix++ {
+		str += fmt.Sprintf(" * %v", e.Operands[ix])
+	}
+	str += " )"
+	return str
+}
+
+func (e exp) String() string {
+	return fmt.Sprintf("exp( %v )", e.Arg)
+}
+
+func (e log) String() string {
+	return fmt.Sprintf("log( %v )", e.Arg)
+}
+
+func (e pow) String() string {
+	return fmt.Sprintf("( %v^%v )", e.Base, e.Exponent)
+}
