+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/victorbrun/gosymbol
 
-go 1.18
-
-require golang.org/x/text v0.3.7 // indirect
+go 1.23
